+++ kaithem/src/tests/test_chandler.py
@@ -1022,6 +1022,7 @@ def test_tag_backtrack_feature():
 
         core.wait_frame()
         core.wait_frame()
+        time.sleep(0.1)
         assert tagpoints.Tag("/test_bt").value == 1
 
         s.cues["default"].set_value_immediate("/test_bt", "value", 2)